move runtime api into separate file
